+++ .gitignore
@@ -129,13 +129,12 @@ dmypy.json
 .pyre/
 *.pdf
 *.mat
-output/fig03_MT_RFs/*.pkl
+*.pkl
 *.ckpt
 *.tfevents.*
 latest_checkpoint.txt
-model.pt
-
-figures/fig02/*.xls
-scripts/*.png
-scripts/*.pt
+*.pt
+*.xls
+*.png
 *.txt
+*.xlsx
