+++ scripts/get_shifters.py
@@ -38,9 +38,8 @@ NUM_WORKERS = int(os.cpu_count() / 2)
 from datasets.pixel import Pixel
 from models.utils import plot_stas
 
-from datasets.pixel.utils import get_stim_list
-sesslist = list(get_stim_list().keys())
-sessname = sesslist[12]
+sesslist = ['20191119', '20191121', '20191205', '20191206',  '20191231', '20200304', '20220601', '20220610']
+sessname = sesslist[7]
 datadir = '/home/jake/Data/Datasets/MitchellV1FreeViewing/stim_movies/'
 NBname = 'shifter_{}'.format(sessname)
 dirname = os.path.join('.', 'checkpoints', NBname)
@@ -49,7 +48,7 @@ print(dirname)
 valid_eye_rad = 9.2
 ds = Pixel(datadir,
     sess_list=[sessname],
-    requested_stims=['Dots', 'Gabor'],
+    requested_stims=['Gabor'],
     num_lags=12,
     downsample_t=2,
     download=True,
@@ -57,7 +56,7 @@ ds = Pixel(datadir,
     ctr=np.array([0,0]),
     fixations_only=True,
     load_shifters=False,
-    spike_sorting='kilowf',
+    spike_sorting='kilo',
     covariate_requests={
         'fixation_onset': {'tent_ctrs': np.arange(-5, 40, 3)},
         'frame_tent': {'ntents': 20}}
@@ -200,10 +199,10 @@ else:
 from models.shifters import ShifterModel
 from models.utils import eval_model
 from NDNT.training import Trainer, EarlyStopping
-from NDNT.utils import NDNutils, load_model
+from NDNT.utils import load_model
 
 input_dims = [1, win_size, win_size] + [ds.num_lags]
-num_filters = [6, 6, 5]
+num_filters = [16, 16, 16]
 filter_width = [7, 5, 3]
 
 data = next(iter(train_dl))
@@ -212,15 +211,16 @@ lambdas = [1e-4, 1e-3, 1e-2]
 for irun in range(5):
     d2x = np.random.randint(len(lambdas))
     d2t = np.random.randint(len(lambdas))
-    center = np.random.randint(len(lambdas))
+    # center = np.random.randint(len(lambdas))
     
     cr0 = ShifterModel(input_dims, NC=ds.NC,
         num_subunits=num_filters,
         filter_width=filter_width,
         cids=cids,
+        scaffold=[2],
         drifter=False,
         noise_sigma=0,
-        reg_vals={'d2x': d2x, 'd2t': d2t, 'center': center},
+        reg_vals={'d2x': d2x, 'd2t': d2t}, #, 'center': center
         reg_hidden=None,
         reg_readout={'l2':1e-6},
         modifiers = {'stimlist': ['frame_tent'],
@@ -252,7 +252,7 @@ for irun in range(5):
 
     earlystopping = EarlyStopping(patience=10, verbose=False)
 
-    trainer = Trainer(cr0, optimizer=optimizer,
+    trainer = Trainer(optimizer=optimizer,
         device = device,
         max_epochs=500,
         verbose=1,
